requirements.txt
defusedxml==0.6.0
Django==2.2.1
django-tastypie==0.14.2
lxml==4.3.3
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
